chore: move to `Application` class package to view
